Migrating to Flask-Mail-SendGrid and API Key authentication.
SendGrid has deprecated basic authentication in favor of API key authentication. On one hand, this change locks
this project template into using SendGrid, but on the other, provides a touch better security out of the gate.

Please feel free to reject this change if you prefer SMTP-based flexibility over an extension locking in to one
provider. I just migrated my own project and it was a low lift to send this your way as well!